Reorganize rules namespaces (#66)
