+++ trunk/Source/CslaContrib.Caliburn.Micro.V2.WPF.Net45/CslaContrib.Caliburn.Micro.V2.WPF.Net45.csproj
@@ -59,6 +59,9 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\GlobalAssemblyInfo.cs">
+      <Link>Properties\GlobalAssemblyInfo.cs</Link>
+    </Compile>
     <Compile Include="EnumerableExtensions.cs" />
     <Compile Include="ErrorEventArgs.cs" />
     <Compile Include="ExecuteEventArgs.cs" />
