+++ trunk/Source/CslaContrib.Net45/CslaContrib.Net45.csproj
@@ -79,8 +79,12 @@
     <Compile Include="Rules\ConditionalRequired\RequiredWhenIsNotNew.cs" />
     <Compile Include="Rules\ConditionalRequired\RequiredWhenNew.cs" />
     <Compile Include="Rules\AuthorizationRules\RestrictByStatusOrIsInRole.cs" />
+    <Compile Include="Rules\ShortCircuitingRules\StopIfAnyAdditionalHasValue.cs" />
     <Compile Include="Rules\ShortCircuitingRules\StopIfIsNew.cs" />
+    <Compile Include="Rules\ShortCircuitingRules\StopIfNotCanWrite.cs" />
     <Compile Include="Rules\ShortCircuitingRules\StopIfNotFieldExists.cs" />
+    <Compile Include="Rules\ShortCircuitingRules\StopIfNotIsExisting.cs" />
+    <Compile Include="Rules\ShortCircuitingRules\StopIfNotIsNew.cs" />
     <Compile Include="SmartDateExtendedParser.cs" />
   </ItemGroup>
   <ItemGroup>
